+++ src/spo2_algorithm.cpp
@@ -2,20 +2,20 @@
 
 #define MAX_FILTER_ORDER 8
 
-void get_oxygen_level(double *ppg_ir, double *ppg_red, int data_size, int sampling_rate, double callib_coef1, double callib_coef2, double callib_coef3, double *oxygen_level)
+void get_oxygen_level(float *ppg_ir, float *ppg_red, int data_size, int sampling_rate, float callib_coef1, float callib_coef2, float callib_coef3, float *oxygen_level)
 {
-    double *red_raw = new double[data_size];
-    double *ir_raw = new double[data_size];
+    float *red_raw = new float[data_size];
+    float *ir_raw = new float[data_size];
     int filter_shift = 25; // to get rif of filtereing artifact, dont use first elements
     int new_size = data_size - filter_shift;
-    double *new_red_raw = red_raw + filter_shift;
-    double *new_ir_raw = ir_raw + filter_shift;
-    memcpy (red_raw, ppg_red, data_size * sizeof (double));
-    memcpy (ir_raw, ppg_ir, data_size * sizeof (double));
+    float *new_red_raw = red_raw + filter_shift;
+    float *new_ir_raw = ir_raw + filter_shift;
+    memcpy (red_raw, ppg_red, data_size * sizeof (float));
+    memcpy (ir_raw, ppg_ir, data_size * sizeof (float));
 
     // need prefiltered mean of red and ir for dc
-    double mean_red = mean (new_red_raw, new_size);
-    double mean_ir = mean (new_ir_raw, new_size);
+    float mean_red = mean (new_red_raw, new_size);
+    float mean_ir = mean (new_ir_raw, new_size);
 
     // filtering(full size)
     detrend (red_raw, data_size, (int)DetrendOperations::CONSTANT);
@@ -24,14 +24,14 @@ void get_oxygen_level(double *ppg_ir, double *ppg_red, int data_size, int sampli
     perform_bandpass (ir_raw, data_size, sampling_rate, 0.7, 1.5, 4, (int)FilterTypes::BUTTERWORTH, 0.0);
 
     // calculate AC & DC components using mean & rms:
-    double redac = rms (new_red_raw, new_size);
-    double irac = rms (new_ir_raw, new_size);
-    double reddc = mean_red;
-    double irdc = mean_ir;
+    float redac = rms (new_red_raw, new_size);
+    float irac = rms (new_ir_raw, new_size);
+    float reddc = mean_red;
+    float irdc = mean_ir;
 
     // https://www.maximintegrated.com/en/design/technical-documents/app-notes/6/6845.html
-    double r = (redac / reddc) / (irac / irdc);
-    double spo2 = callib_coef1 * r * r + callib_coef2 * r + callib_coef3;
+    float r = (redac / reddc) / (irac / irdc);
+    float spo2 = callib_coef1 * r * r + callib_coef2 * r + callib_coef3;
     if (spo2 > 100.0)
     {
         spo2 = 100.0;
@@ -46,11 +46,11 @@ void get_oxygen_level(double *ppg_ir, double *ppg_red, int data_size, int sampli
     delete[] ir_raw;
 }
 
-void detrend (double *data, int data_len, int detrend_operation)
+void detrend (float *data, int data_len, int detrend_operation)
 {
     if (detrend_operation == (int)DetrendOperations::CONSTANT)
     {
-        double mean = 0.0;
+        float mean = 0.0;
         // subtract mean from data
         for (int i = 0; i < data_len; i++)
         {
@@ -65,24 +65,24 @@ void detrend (double *data, int data_len, int detrend_operation)
     if (detrend_operation == (int)DetrendOperations::LINEAR)
     {
         // use mean and gradient to find a line
-        double mean_x = (data_len - 1) / 2.0;
-        double mean_y = 0;
+        float mean_x = (data_len - 1) / 2.0;
+        float mean_y = 0;
         for (int i = 0; i < data_len; i++)
         {
             mean_y += data[i];
         }
         mean_y /= data_len;
-        double temp_xy = 0.0;
-        double temp_xx = 0.0;
+        float temp_xy = 0.0;
+        float temp_xx = 0.0;
         for (int i = 0; i < data_len; i++)
         {
             temp_xy += i * data[i];
             temp_xx += i * i;
         }
-        double s_xy = temp_xy / data_len - mean_x * mean_y;
-        double s_xx = temp_xx / data_len - mean_x * mean_x;
-        double grad = s_xy / s_xx;
-        double y_int = mean_y - grad * mean_x;
+        float s_xy = temp_xy / data_len - mean_x * mean_y;
+        float s_xx = temp_xx / data_len - mean_x * mean_x;
+        float grad = s_xy / s_xx;
+        float y_int = mean_y - grad * mean_x;
         for (int i = 0; i < data_len; i++)
         {
             data[i] = data[i] - (grad * i + y_int);
@@ -90,12 +90,12 @@ void detrend (double *data, int data_len, int detrend_operation)
     }
 }
 
-void perform_bandpass (double *data, int data_len, int sampling_rate, double start_freq, double stop_freq, int order, int filter_type, double ripple)
+void perform_bandpass (float *data, int data_len, int sampling_rate, float start_freq, float stop_freq, int order, int filter_type, float ripple)
 {
-    double center_freq = (start_freq + stop_freq) / 2.0;
-    double band_width = stop_freq - start_freq;
+    float center_freq = (start_freq + stop_freq) / 2.0;
+    float band_width = stop_freq - start_freq;
     Dsp::Filter *f = NULL;
-    double *filter_data[1];
+    float *filter_data[1];
     filter_data[0] = data;
 
     switch (static_cast<FilterTypes> (filter_type))
